+++ packages/cdai/src/component_helpers/StorybookHelper.js
@@ -1,12 +1,17 @@
 //
-// Copyright IBM Corp. 2020, 2020
+// Copyright IBM Corp. 2020, 2021
 //
 // This source code is licensed under the Apache-2.0 license found in the
 // LICENSE file in the root directory of this source tree.
 //
 
-import { sectionTitle } from '../../config';
+import { getPathForComponent } from '../../../core/story-structure';
 
-export const getComponentLabel = (component) => {
-  return `${sectionTitle}/${component}`;
+export const getComponentLabel = (componentName) => {
+  const title =
+    // if the component isn't in the master structure, put it in a lost+found section
+    getPathForComponent('a', componentName) ||
+    `Cloud & Cognitive/Lost + found/${componentName}`;
+
+  return title;
 };